requirements.txt
absl-py==2.1.0
adagio==0.2.6
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
arch==7.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==25.1.0
autots==0.6.19
avro==1.11.3
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
boto3==1.36.19
botocore==1.36.19
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmdstanpy==1.2.5
colorama==0.4.6
colorlog==6.9.0
comet_ml==3.49.0
comm==0.2.2
configobj==5.0.9
confluent-kafka==2.3.0
contourpy==1.3.1
coreforecast==0.0.15
cramjam==2.9.1
cryptography==44.0.1
cycler==0.12.1
Cython==3.0.12
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
dtaidistance==2.3.12
dtw-python==1.5.3
dulwich==0.22.7
everett==3.1.0
exceptiongroup==1.2.2
executing==2.2.0
fastavro==1.8.4
fastjsonschema==2.21.1
fastparquet==2024.11.0
filterpy==1.4.5
fire==0.7.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fs==2.4.16
fsspec==2025.2.0
fugue==0.9.1
furl==2.1.3
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
gluonts==0.16.0
google-api-core==2.24.2
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.70.0
gspread==6.2.0
h11==0.14.0
h5py==3.12.1
hmmlearn==0.3.3
holidays==0.66
hopsworks==4.1.8
hopsworks-aiomysql==0.2.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
javaobj-py3==0.4.4
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.8.0
keras-self-attention==0.51.0
kiwisolver==1.4.8
lab==8.4
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
meteostat==1.6.8
mistune==3.1.1
ml-dtypes==0.3.2
mne==1.9.0
mock==5.1.0
namex==0.0.8
narwhals==1.26.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
nfoursid==1.0.1
notebook==7.3.2
notebook_shim==0.2.4
numba==0.61.0
numpy==1.26.4
oauthlib==3.2.2
opensearch-py==2.4.2
opt_einsum==3.4.0
optree==0.14.0
optuna==4.2.0
orderedmultidict==1.0.1
overrides==7.7.0
packaging==24.2
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pmdarima==2.0.4
polars==1.22.0
pooch==1.8.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
prophet==1.1.6
proto-plus==1.26.1
protobuf==4.25.6
psutil==6.1.1
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycatch22==0.4.5
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyhumps==1.6.1
pyjks==20.0.0
PyMySQL==1.1.1
pyod==2.0.3
pyparsing==3.2.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pyts==0.13.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.11.2
scikit-base==0.12.0
scikit-learn==1.6.1
scikit-optimize==0.10.2
scikit-posthocs==0.11.2
scipy==1.15.1
seaborn==0.13.2
seasonal==0.3.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shap==0.46.0
simplejson==3.19.3
six==1.17.0
skforecast==0.14.0
skpro==2.9.0
sktime==0.36.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.29
stack-data==0.6.3
stanio==0.5.1
statsforecast==2.0.0
statsmodels==0.14.4
stumpy==1.13.0
ta==0.11.0
tbats==1.1.3
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tomli==2.2.1
toolz==0.12.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triad==0.9.8
tsfresh==0.20.3
tslearn==0.6.3
twofish==0.3.0
txt2tags==3.9
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
u8darts==0.31.0
uri-template==1.3.0
urllib3==2.3.0
utilsforecast==0.2.11
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
wurlitzer==3.1.1
xarray==2025.1.2
xgboost==2.1.4
zipp==3.21.0
